+++ Arduino-project/src/components/interfaces/InterfaceLineSensor.h
@@ -4,7 +4,7 @@
 class ILineSensor {
     public:
         virtual void init() = 0;
-        virtual int getLinePosition() = 0;
+        virtual double getLinePosition() = 0;
 };
 
 #endif // ILINESENSOR_H
